==> scripts/1-OCR with Pytesseract (Jupyter Notebook).ipynb <==
# Install Packages

!pip install pytesseract pillow
from PIL import Image
import pytesseract
import os
import glob

## For Windows computer only
#Set tesseract path (not needed on Ubuntu once pytesseract is installed in command line)
pytesseract.pytesseract.tesseract_cmd = r'C:\Users\tuk51061\AppData\Local\Programs\Tesseract-OCR\tesseract'

# OCR for Multiple Books in Subfolders

def ocr_all_files(folder_path):
    text = ""
    tiffs_path = os.path.join(folder_path, "TIFFs to OCR")
    
    if os.path.exists(tiffs_path):
        tiff_files = [file for file in sorted(os.listdir(tiffs_path)) if file.lower().endswith('.tif')]
        for file in tiff_files:
            file_path = os.path.join(tiffs_path, file)
            
            try:
                image = Image.open(file_path)
                img_text = pytesseract.image_to_string(image, lang='eng')     
                text += img_text + '\n'
            except Exception as e:
                print("Skipped:", file_path, " - Unidentified image format")
    return text
    
main_directory = '\PATHNAME'
output_directory = '\PATHNAME'

for book_subdir in os.listdir(main_directory):
    book_subdir_path = os.path.join(main_directory, book_subdir)
    if os.path.isdir(book_subdir_path):
        output_text_file = os.path.join(output_directory, book_subdir + '.txt')
        extracted_text = ocr_all_files(book_subdir_path)
        
        with open(output_text_file, 'w', encoding='utf-8') as f:
            f.write(extracted_text)
            print("OCR completed for", book_subdir, "and text saved to", output_text_file)
